+++ sqlalchemy_redshift/__init__.py
@@ -21,7 +21,15 @@ __version__ = get_distribution('sqlalchemy-redshift').version
 
 from sqlalchemy.dialects import registry
 
-registry.register("redshift", "sqlalchemy_redshift.dialect", "RedshiftDialect")
 registry.register(
-    "redshift.psycopg2", "sqlalchemy_redshift.dialect", "RedshiftDialect"
+    "redshift", "sqlalchemy_redshift.dialect",
+    "RedshiftDialect_psycopg2"
+)
+registry.register(
+    "redshift.psycopg2", "sqlalchemy_redshift.dialect",
+    "RedshiftDialect_psycopg2"
+)
+registry.register(
+    'redshift+psycopg2cffi', 'sqlalchemy_redshift.dialect',
+    'RedshiftDialect_psycopg2cffi',
 )