+++ tests/test_compiler.py
@@ -1,9 +1,8 @@
 from sqlalchemy import func, select
-from sqlalchemy_redshift.dialect import RedshiftDialect
 
 
-def test_func_now():
-    dialect = RedshiftDialect()
+def test_func_now(stub_redshift_dialect):
+    dialect = stub_redshift_dialect
     s = select([func.NOW().label("time")])
     compiled = s.compile(dialect=dialect)
     assert str(compiled) == "SELECT SYSDATE AS time"