+++ tests/test_default_ssl.py
@@ -5,8 +5,8 @@ from pkg_resources import resource_filename
 CERT_PATH = resource_filename("sqlalchemy_redshift", "redshift-ca-bundle.crt")
 
 
-def test_ssl_args():
-    engine = sa.create_engine('redshift+psycopg2://test')
+def test_ssl_args(redshift_dialect_flavor):
+    engine = sa.create_engine('{}://test'.format(redshift_dialect_flavor))
     dialect = engine.dialect
     url = engine.url