+++ tests/test_materialized_views.py
@@ -14,7 +14,7 @@ def selectable():
     return select([table.c.id, table.c.name], from_obj=table)
 
 
-def test_basic_materialized_view(selectable):
+def test_basic_materialized_view(selectable, stub_redshift_dialect):
     expected_result = """
     CREATE MATERIALIZED VIEW test_view
     AS SELECT t1.id, t1.name FROM t1
@@ -23,10 +23,11 @@ def test_basic_materialized_view(selectable):
         "test_view",
         selectable
     )
-    assert clean(expected_result) == clean(compile_query(view))
+    assert clean(expected_result) == \
+        clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_no_backup_materialized_view(selectable):
+def test_no_backup_materialized_view(selectable, stub_redshift_dialect):
     expected_result = """
     CREATE MATERIALIZED VIEW test_view
     BACKUP NO
@@ -37,10 +38,11 @@ def test_no_backup_materialized_view(selectable):
         selectable,
         backup=False
     )
-    assert clean(expected_result) == clean(compile_query(view))
+    assert clean(expected_result) == \
+        clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_diststyle_materialized_view(selectable):
+def test_diststyle_materialized_view(selectable, stub_redshift_dialect):
     expected_result = """
     CREATE MATERIALIZED VIEW test_view
     DISTSTYLE ALL
@@ -51,10 +53,11 @@ def test_diststyle_materialized_view(selectable):
         selectable,
         diststyle='ALL'
     )
-    assert clean(expected_result) == clean(compile_query(view))
+    assert clean(expected_result) == \
+        clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_distkey_materialized_view(selectable):
+def test_distkey_materialized_view(selectable, stub_redshift_dialect):
     expected_result = """
     CREATE MATERIALIZED VIEW test_view
     DISTKEY (id)
@@ -66,10 +69,11 @@ def test_distkey_materialized_view(selectable):
             selectable,
             distkey=key
         )
-        assert clean(expected_result) == clean(compile_query(view))
+        assert clean(expected_result) == \
+            clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_sortkey_materialized_view(selectable):
+def test_sortkey_materialized_view(selectable, stub_redshift_dialect):
     expected_result = """
     CREATE MATERIALIZED VIEW test_view
     SORTKEY (id)
@@ -81,10 +85,13 @@ def test_sortkey_materialized_view(selectable):
             selectable,
             sortkey=key
         )
-        assert clean(expected_result) == clean(compile_query(view))
+        assert clean(expected_result) == \
+            clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_interleaved_sortkey_materialized_view(selectable):
+def test_interleaved_sortkey_materialized_view(
+        selectable, stub_redshift_dialect
+):
     expected_result = """
     CREATE MATERIALIZED VIEW test_view
     INTERLEAVED SORTKEY (id)
@@ -96,28 +103,33 @@ def test_interleaved_sortkey_materialized_view(selectable):
             selectable,
             interleaved_sortkey=key
         )
-        assert clean(expected_result) == clean(compile_query(view))
+        assert clean(expected_result) == \
+            clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_drop_materialized_view():
+def test_drop_materialized_view(stub_redshift_dialect):
     expected_result = "DROP MATERIALIZED VIEW test_view"
     view = dialect.DropMaterializedView("test_view")
-    assert clean(expected_result) == clean(compile_query(view))
+    assert clean(expected_result) == \
+        clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_drop_materialized_view_if_exists():
+def test_drop_materialized_view_if_exists(stub_redshift_dialect):
     expected_result = "DROP MATERIALIZED VIEW IF EXISTS test_view"
     view = dialect.DropMaterializedView("test_view", if_exists=True)
-    assert clean(expected_result) == clean(compile_query(view))
+    assert clean(expected_result) == \
+        clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_drop_materialized_view_cascade():
+def test_drop_materialized_view_cascade(stub_redshift_dialect):
     expected_result = "DROP MATERIALIZED VIEW test_view CASCADE"
     view = dialect.DropMaterializedView("test_view", cascade=True)
-    assert clean(expected_result) == clean(compile_query(view))
+    assert clean(expected_result) == \
+        clean(compile_query(view, stub_redshift_dialect))
 
 
-def test_refresh_materialized_view():
+def test_refresh_materialized_view(stub_redshift_dialect):
     expected_result = "REFRESH MATERIALIZED VIEW test_view"
     view = dialect.RefreshMaterializedView("test_view")
-    assert clean(expected_result) == clean(compile_query(view))
+    assert clean(expected_result) == \
+        clean(compile_query(view, stub_redshift_dialect))