+++ tests/test_reflection.py
@@ -3,14 +3,12 @@ from sqlalchemy import MetaData, Table
 from sqlalchemy.schema import CreateTable
 from sqlalchemy.exc import NoSuchTableError
 
-from sqlalchemy_redshift import dialect
-
 from rs_sqla_test_utils import models, utils
 
 
-def table_to_ddl(table):
+def table_to_ddl(table, _dialect):
     return str(CreateTable(table).compile(
-        dialect=dialect.RedshiftDialect()
+        dialect=_dialect
     ))
 
 
@@ -165,18 +163,19 @@ models_and_ddls = [
 
 
 @pytest.mark.parametrize("model, ddl", models_and_ddls)
-def test_definition(model, ddl):
-    model_ddl = table_to_ddl(model.__table__)
+def test_definition(model, ddl, stub_redshift_dialect):
+    model_ddl = table_to_ddl(model.__table__, stub_redshift_dialect)
     assert utils.clean(model_ddl) == utils.clean(ddl)
 
 
 @pytest.mark.parametrize("model, ddl", models_and_ddls)
 def test_reflection(redshift_session, model, ddl):
+    _dialect = redshift_session.bind.dialect
     metadata = MetaData(bind=redshift_session.bind)
     schema = model.__table__.schema
     table = Table(model.__tablename__, metadata,
                   schema=schema, autoload=True)
-    introspected_ddl = table_to_ddl(table)
+    introspected_ddl = table_to_ddl(table, _dialect)
     assert utils.clean(introspected_ddl) == utils.clean(ddl)